requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
# AWS SDK for S3
boto3 
# Database ORM and Driver
sqlalchemy
psycopg2-binary
# For QR code generation logic
qrcode[pil]

passlib
bcrypt>=4.0,<5.0
python-jose
python-multipart
reportlab
psycopg2-binary
pillow 
boto3